+++ moor/test/data/tables/custom_tables.g.dart
@@ -11,7 +11,12 @@ class Config extends DataClass implements Insertable<Config> {
   final String configKey;
   final String configValue;
   final SyncType syncState;
-  Config({@required this.configKey, this.configValue, this.syncState});
+  final SyncType syncStateImplicit;
+  Config(
+      {@required this.configKey,
+      this.configValue,
+      this.syncState,
+      this.syncStateImplicit});
   factory Config.fromData(Map<String, dynamic> data, GeneratedDatabase db,
       {String prefix}) {
     final effectivePrefix = prefix ?? '';
@@ -24,6 +29,9 @@ class Config extends DataClass implements Insertable<Config> {
           .mapFromDatabaseResponse(data['${effectivePrefix}config_value']),
       syncState: ConfigTable.$converter0.mapToDart(intType
           .mapFromDatabaseResponse(data['${effectivePrefix}sync_state'])),
+      syncStateImplicit: ConfigTable.$converter1.mapToDart(
+          intType.mapFromDatabaseResponse(
+              data['${effectivePrefix}sync_state_implicit'])),
     );
   }
   @override
@@ -39,6 +47,11 @@ class Config extends DataClass implements Insertable<Config> {
       final converter = ConfigTable.$converter0;
       map['sync_state'] = Variable<int>(converter.mapToSql(syncState));
     }
+    if (!nullToAbsent || syncStateImplicit != null) {
+      final converter = ConfigTable.$converter1;
+      map['sync_state_implicit'] =
+          Variable<int>(converter.mapToSql(syncStateImplicit));
+    }
     return map;
   }
 
@@ -53,6 +66,9 @@ class Config extends DataClass implements Insertable<Config> {
       syncState: syncState == null && nullToAbsent
           ? const Value.absent()
           : Value(syncState),
+      syncStateImplicit: syncStateImplicit == null && nullToAbsent
+          ? const Value.absent()
+          : Value(syncStateImplicit),
     );
   }
 
@@ -63,6 +79,8 @@ class Config extends DataClass implements Insertable<Config> {
       configKey: serializer.fromJson<String>(json['config_key']),
       configValue: serializer.fromJson<String>(json['config_value']),
       syncState: serializer.fromJson<SyncType>(json['sync_state']),
+      syncStateImplicit:
+          serializer.fromJson<SyncType>(json['sync_state_implicit']),
     );
   }
   factory Config.fromJsonString(String encodedJson,
@@ -76,71 +94,88 @@ class Config extends DataClass implements Insertable<Config> {
       'config_key': serializer.toJson<String>(configKey),
       'config_value': serializer.toJson<String>(configValue),
       'sync_state': serializer.toJson<SyncType>(syncState),
+      'sync_state_implicit': serializer.toJson<SyncType>(syncStateImplicit),
     };
   }
 
-  Config copyWith({String configKey, String configValue, SyncType syncState}) =>
+  Config copyWith(
+          {String configKey,
+          String configValue,
+          SyncType syncState,
+          SyncType syncStateImplicit}) =>
       Config(
         configKey: configKey ?? this.configKey,
         configValue: configValue ?? this.configValue,
         syncState: syncState ?? this.syncState,
+        syncStateImplicit: syncStateImplicit ?? this.syncStateImplicit,
       );
   @override
   String toString() {
     return (StringBuffer('Config(')
           ..write('configKey: $configKey, ')
           ..write('configValue: $configValue, ')
-          ..write('syncState: $syncState')
+          ..write('syncState: $syncState, ')
+          ..write('syncStateImplicit: $syncStateImplicit')
           ..write(')'))
         .toString();
   }
 
   @override
   int get hashCode => $mrjf($mrjc(
-      configKey.hashCode, $mrjc(configValue.hashCode, syncState.hashCode)));
+      configKey.hashCode,
+      $mrjc(configValue.hashCode,
+          $mrjc(syncState.hashCode, syncStateImplicit.hashCode))));
   @override
   bool operator ==(dynamic other) =>
       identical(this, other) ||
       (other is Config &&
           other.configKey == this.configKey &&
           other.configValue == this.configValue &&
-          other.syncState == this.syncState);
+          other.syncState == this.syncState &&
+          other.syncStateImplicit == this.syncStateImplicit);
 }
 
 class ConfigCompanion extends UpdateCompanion<Config> {
   final Value<String> configKey;
   final Value<String> configValue;
   final Value<SyncType> syncState;
+  final Value<SyncType> syncStateImplicit;
   const ConfigCompanion({
     this.configKey = const Value.absent(),
     this.configValue = const Value.absent(),
     this.syncState = const Value.absent(),
+    this.syncStateImplicit = const Value.absent(),
   });
   ConfigCompanion.insert({
     @required String configKey,
     this.configValue = const Value.absent(),
     this.syncState = const Value.absent(),
+    this.syncStateImplicit = const Value.absent(),
   }) : configKey = Value(configKey);
   static Insertable<Config> custom({
     Expression<String> configKey,
     Expression<String> configValue,
     Expression<int> syncState,
+    Expression<int> syncStateImplicit,
   }) {
     return RawValuesInsertable({
       if (configKey != null) 'config_key': configKey,
       if (configValue != null) 'config_value': configValue,
       if (syncState != null) 'sync_state': syncState,
+      if (syncStateImplicit != null) 'sync_state_implicit': syncStateImplicit,
     });
   }
 
   ConfigCompanion copyWith(
       {Value<String> configKey,
       Value<String> configValue,
-      Value<SyncType> syncState}) {
+      Value<SyncType> syncState,
+      Value<SyncType> syncStateImplicit}) {
     return ConfigCompanion(
       configKey: configKey ?? this.configKey,
       configValue: configValue ?? this.configValue,
       syncState: syncState ?? this.syncState,
+      syncStateImplicit: syncStateImplicit ?? this.syncStateImplicit,
     );
   }
 
@@ -156,6 +191,11 @@ class ConfigCompanion extends UpdateCompanion<Config> {
     if (syncState.present) {
       final converter = ConfigTable.$converter0;
       map['sync_state'] = Variable<int>(converter.mapToSql(syncState.value));
+    }
+    if (syncStateImplicit.present) {
+      final converter = ConfigTable.$converter1;
+      map['sync_state_implicit'] =
+          Variable<int>(converter.mapToSql(syncStateImplicit.value));
     }
     return map;
   }
@@ -191,8 +231,19 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
         $customConstraints: '');
   }
 
+  final VerificationMeta _syncStateImplicitMeta =
+      const VerificationMeta('syncStateImplicit');
+  GeneratedIntColumn _syncStateImplicit;
+  GeneratedIntColumn get syncStateImplicit =>
+      _syncStateImplicit ??= _constructSyncStateImplicit();
+  GeneratedIntColumn _constructSyncStateImplicit() {
+    return GeneratedIntColumn('sync_state_implicit', $tableName, true,
+        $customConstraints: '');
+  }
+
   @override
-  List<GeneratedColumn> get $columns => [configKey, configValue, syncState];
+  List<GeneratedColumn> get $columns =>
+      [configKey, configValue, syncState, syncStateImplicit];
   @override
   ConfigTable get asDslTable => this;
   @override
@@ -217,6 +268,7 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
               data['config_value'], _configValueMeta));
     }
     context.handle(_syncStateMeta, const VerificationResult.success());
+    context.handle(_syncStateImplicitMeta, const VerificationResult.success());
     return context;
   }
 
@@ -234,6 +286,8 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
   }
 
   static TypeConverter<SyncType, int> $converter0 = const SyncTypeConverter();
+  static TypeConverter<SyncType, int> $converter1 =
+      const EnumIndexConverter<SyncType>(SyncType.values);
   @override
   bool get dontWriteConstraints => true;
 }
@@ -1300,6 +1354,8 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
       configKey: row.readString('config_key'),
       configValue: row.readString('config_value'),
       syncState: ConfigTable.$converter0.mapToDart(row.readInt('sync_state')),
+      syncStateImplicit:
+          ConfigTable.$converter1.mapToDart(row.readInt('sync_state_implicit')),
     );
   }
 
@@ -1384,6 +1440,8 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
       configKey: row.readString('config_key'),
       configValue: row.readString('config_value'),
       syncState: ConfigTable.$converter0.mapToDart(row.readInt('sync_state')),
+      syncStateImplicit:
+          ConfigTable.$converter1.mapToDart(row.readInt('sync_state_implicit')),
     );
   }
 
@@ -1481,17 +1539,21 @@ class ReadRowIdResult {
   final String configKey;
   final String configValue;
   final SyncType syncState;
+  final SyncType syncStateImplicit;
   ReadRowIdResult({
     this.rowid,
     this.configKey,
     this.configValue,
     this.syncState,
+    this.syncStateImplicit,
   });
   @override
   int get hashCode => $mrjf($mrjc(
       rowid.hashCode,
-      $mrjc(configKey.hashCode,
-          $mrjc(configValue.hashCode, syncState.hashCode))));
+      $mrjc(
+          configKey.hashCode,
+          $mrjc(configValue.hashCode,
+              $mrjc(syncState.hashCode, syncStateImplicit.hashCode)))));
   @override
   bool operator ==(dynamic other) =>
       identical(this, other) ||
@@ -1499,5 +1561,6 @@ class ReadRowIdResult {
           other.rowid == this.rowid &&
           other.configKey == this.configKey &&
           other.configValue == this.configValue &&
-          other.syncState == this.syncState);
+          other.syncState == this.syncState &&
+          other.syncStateImplicit == this.syncStateImplicit);
 }