+++ moor_generator/lib/src/analyzer/moor/parser.dart
@@ -27,7 +27,8 @@ class MoorParser {
         final importStmt = parsedStmt;
         importStatements.add(importStmt);
       } else if (parsedStmt is TableInducingStatement) {
-        createdReaders.add(CreateTableReader(parsedStmt, step));
+        createdReaders
+            .add(CreateTableReader(parsedStmt, step, importStatements));
       } else if (parsedStmt is CreateTriggerStatement) {
         // the table will be resolved in the analysis step
         createdEntities.add(MoorTrigger.fromMoor(parsedStmt, step.file));